// src/types/enums/GameSettingsActionEnum.ts
export enum GameSettingsActionEnum {
  UPDATE_BOARD_WIDTH = 'UPDATE_BOARD_WIDTH',
  UPDATE_BOARD_HEIGHT = 'UPDATE_BOARD_HEIGHT',
  UPDATE_ROW_COUNT = 'UPDATE_ROW_COUNT',
  UPDATE_COL_COUNT = 'UPDATE_COL_COUNT',
  UPDATE_SPEED = 'UPDATE_SPEED',
  UPDATE_RADIUS = 'UPDATE_RADIUS',
  RESET_SETTINGS = 'RESET_SETTINGS',
}